+++ packages/web-app-files/src/helpers/resources.ts
@@ -89,6 +89,9 @@ function addSharedWithToShares(shares) {
       previousShare?.file_source === share.file_source
     ) {
       if (ShareTypes.containsAnyValue(ShareTypes.authenticated, [parseInt(share.share_type)])) {
+        if (share.stime > previousShare.stime) {
+          previousShare.stime = share.stime
+        }
         previousShare.sharedWith.push({
           username: share.share_with,
           name: share.share_with_displayname,