+++ tests/e2e/cucumber/steps/ui/resources.ts
@@ -224,17 +224,23 @@ Then(
     }
   }
 )
+
 When(
-  '{string} searches {string} using the global search',
-  async function (this: World, stepUser: string, keyword: string): Promise<void> {
+  /^"([^"]*)" searches "([^"]*)" using the global search( and presses enter)?$/,
+  async function (this: World, stepUser: string, keyword: string, command: string): Promise<void> {
+    keyword = keyword ?? ''
+    const pressEnter = !!command && command.endsWith('presses enter')
     const { page } = this.actorsEnvironment.getActor({ key: stepUser })
     const resourceObject = new objects.applicationFiles.Resource({ page })
-    await resourceObject.searchResource({ keyword })
+    await resourceObject.searchResource({
+      keyword,
+      pressEnter
+    })
   }
 )
 
 Then(
-  /^following resources (should|should not) be displayed in the (search list|files list) for user "([^"]*)"?$/,
+  /^following resources (should|should not) be displayed in the (search list|files list) for user "([^"]*)"$/,
   async function (
     this: World,
     actionType: string,